메인 헬스장 입장 QR생성

+++ gym/src/main/java/com/gym/gym/domain/QRcode.java
@@ -10,7 +10,7 @@ import lombok.Data;
 @Alias("QRcode")
 public class QRcode {
     
-    private int no;
+    private long no;
     private int userNo;
     private String uuid;
     private Date createdAt;
